Agrego etiquetas de condiciones a User, Wine y cambio tipo atributo en Ticket, TicketDetail, Accesory

+++ src/main/java/com/grupo7/vinoteca/entities/TicketDetail.java
@@ -15,9 +15,9 @@ import javax.persistence.*;
 @NoArgsConstructor
 public class TicketDetail extends Base {
 
-    private int price;
+    private long price;
     private int stock;
-    private int subtotal;
+    private long subtotal;
 
     @OneToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "fk_wine")
